ACC-2017 WIP Change from testing java datatypes to postgresql datatypes

+++ contentgrid-appserver-query-engine/src/main/java/com/contentgrid/appserver/query/engine/JOOQTableCreator.java
@@ -140,7 +140,7 @@ public class JOOQTableCreator {
             case UUID -> SQLDataType.UUID;
             case TEXT -> SQLDataType.CLOB;
             case LONG -> SQLDataType.BIGINT;
-            case DOUBLE -> SQLDataType.DOUBLE;
+            case DOUBLE -> SQLDataType.DECIMAL;
             case BOOLEAN -> SQLDataType.BOOLEAN;
             case DATETIME -> SQLDataType.INSTANT;
         };
